refactor(layout): split scrolling's read-back half off (#966)
`ScrollingLayout.swift` reached 347 lines against the 350
ceiling, so it splits before it crosses (AGENTS.md §2.1) rather
than after. `viewportRest` and `rowOverflows` move to
`ScrollingLayout+Viewport.swift` — what a caller asks the layout
ABOUT a row without materializing its frames, both derived from
the same `metrics` the frames come from. No behavior change.
